add a public method to register new DataType in a plugin
